Move downtime api to new permissions model

+++ bergamot-ui/src/main/java/com/intrbiz/bergamot/ui/api/DowntimeAPIRouter.java
@@ -6,9 +6,7 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import com.intrbiz.Util;
 import com.intrbiz.balsa.engine.route.Router;
-import com.intrbiz.balsa.error.http.BalsaNotFound;
 import com.intrbiz.balsa.metadata.WithDataAdapter;
 import com.intrbiz.bergamot.data.BergamotDB;
 import com.intrbiz.bergamot.metadata.IsaObjectId;
@@ -25,7 +23,6 @@ import com.intrbiz.metadata.IsaInt;
 import com.intrbiz.metadata.JSON;
 import com.intrbiz.metadata.Param;
 import com.intrbiz.metadata.Prefix;
-import com.intrbiz.metadata.RequirePermission;
 import com.intrbiz.metadata.RequireValidPrincipal;
 
 @Prefix("/api/downtime")
@@ -35,26 +32,30 @@ public class DowntimeAPIRouter extends Router<BergamotApp>
 
     @Get("/id/:id")
     @JSON(notFoundIfNull = true)
-    @RequirePermission("api.read.downtime")
     @WithDataAdapter(BergamotDB.class)
     public DowntimeMO getComment(BergamotDB db, @IsaObjectId(session = false) UUID id)
     {
-        return Util.nullable(db.getDowntime(id), Downtime::toMO);
+        Downtime downtime = notNull(db.getDowntime(id));
+        require(permission("read.downtime", downtime.getCheckId()));
+        return downtime.toMO();
     }
     
     @Get("/id/:id/remove")
     @JSON()
-    @RequirePermission("api.write.downtime.remove")
     @WithDataAdapter(BergamotDB.class)
     public Boolean removeDowntime(BergamotDB db, @IsaObjectId(session = false) UUID id)
     {
-        db.removeDowntime(id);
+        Downtime downtime = db.getDowntime(id);
+        if (downtime != null)
+        {
+            require(permission("remove.downtime", downtime.getCheckId()));
+            db.removeDowntime(id);
+        }
         return true;
     }
     
     @Get("/for-object/id/:id")
     @JSON()
-    @RequirePermission("api.read.downtime")
     @WithDataAdapter(BergamotDB.class)
     public List<DowntimeMO> getDowntimeForObject(
             BergamotDB db, 
@@ -63,12 +64,12 @@ public class DowntimeAPIRouter extends Router<BergamotApp>
             @Param("future") @IsaInt(min = 0, max = 365, mandatory = true, defaultValue = 7, coalesce = CoalesceMode.ON_NULL) Integer futureDays
     )
     {
+        require(permission("read.downtime", id));
         return db.getDowntimesForCheck(id, pastDays + " days", futureDays + " days").stream().map(Downtime::toMO).collect(Collectors.toList());
     }
     
     @Any("/add-downtime-to-check/id/:id")
     @JSON()
-    @RequirePermission("api.write.downtime.add")
     @WithDataAdapter(BergamotDB.class)
     public DowntimeMO addDowntimeToCheck(
             BergamotDB db, 
@@ -80,8 +81,8 @@ public class DowntimeAPIRouter extends Router<BergamotApp>
     )
     {
         // TODO: timezone handling
-        Check<?, ?> check = db.getCheck(id);
-        if (check == null) throw new BalsaNotFound("No check with id: " + id);
+        Check<?, ?> check = notNull(db.getCheck(id));
+        require(permission("write.downtime", check));
         //
         Timestamp starts = new Timestamp(startTime.getTime());
         Timestamp ends = new Timestamp(endTime.getTime());
@@ -93,10 +94,11 @@ public class DowntimeAPIRouter extends Router<BergamotApp>
     
     @Get("/id/:id/render")
     @JSON()
-    @RequirePermission("api.read.downtime")
     @WithDataAdapter(BergamotDB.class)
     public String renderDowntime(BergamotDB db, @IsaObjectId(session = false) UUID id)
     {
-        return var("downtime", db.getDowntime(id)) == null ? null : encodeBuffered("include/downtime");
+        Downtime downtime = var("downtime", notNull(db.getDowntime(id)));
+        require(permission("read.downtime", downtime.getCheckId()));
+        return encodeBuffered("include/downtime");
     }
 }
